requirements.txt
-i https://pypi.org/simple
blinker==1.4
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dictdiffer==0.8.1
django-anymail[mailgun]==7.1.0
django-cacheops==4.2
django-cors-headers==3.2.1
django-countries==6.1.2
django-environ==0.4.5
django-filter==2.2.0
django-model-utils==4.0.0
django-request-logging==0.7.1
django-reversion==3.0.7
django-split-settings==1.0.0
django-storages==1.9.1
django-taggit-serializer==0.1.7
django-taggit==1.2.0
django-templated-mail==1.1.1
django-tinymce==3.0.1
django==2.2.10
djangorestframework-camel-case==1.1.2
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
djangorestframework==3.9.1
djoser==1.7
drf-nested-routers==0.91
funcy==1.14
google-api-core[grpc]==1.17.0
google-auth==1.14.0
google-cloud-core==1.3.0
google-cloud-datastore==1.12.0
google-cloud-logging==1.15.0
google-cloud-storage==1.27.0
google-cloud-tasks==1.5.0
google-resumable-media==0.5.0
googleapis-common-protos[grpc]==1.51.0
googlemaps==4.2.0
grpc-google-iam-v1==0.12.3
grpcio==1.28.1
idna==2.9
itypes==1.1.0
jinja2==3.0.0a1
loguru==0.4.1
mangopaysdk==3.7.0
markdown==3.2.1
markupsafe==2.0.0a1
pillow==7.1.1
protobuf==3.11.3
psycopg2-binary==2.8.5
pyasn1-modules==0.2.8
pyasn1==0.4.8
pygments==2.6.1
pyjwt==1.7.1
python-dateutil==2.8.1
pytz==2019.3
redis==3.4.1
requests==2.23.0
retrying==1.3.3
rsa==4.0
sentry-sdk==0.14.3
simplejson==3.17.0
six==1.14.0
sqlparse==0.3.1
trans==2.1.0
uritemplate==3.0.1
urllib3==1.25.8
